+++ src/main/java/com/poly/demo/repository/RoomRepository.java
@@ -1,26 +1,23 @@
 package com.poly.demo.repository;
 
-import com.poly.demo.entity.Branch;
-import com.poly.demo.entity.Room;
-import com.poly.demo.entity.Seat;
-import com.poly.demo.entity.Showtime;
-import com.poly.demo.entity.Ticket;
-import com.poly.demo.entity.TicketFood;
-import com.poly.demo.entity.TicketFoodId;
-import com.poly.demo.entity.User;
+import java.util.List;
+import java.util.Optional;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
+import com.poly.demo.entity.Branch;
+import com.poly.demo.entity.Room;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Long> {
-    List<Room> findByBranch(Branch branch);
-    Optional<Room> findByRoomId(Integer id);
-    void deleteByRoomId(Integer id);
-    Page<Room> findAll(Pageable pageable);
+public interface RoomRepository extends JpaRepository<Room, Integer> {
+	List<Room> findByBranch(Branch branch);
+
+	Optional<Room> findByRoomId(Integer id);
+
+	void deleteByRoomId(Integer id);
+
+	Page<Room> findAll(Pageable pageable);
 }
