+++ src/main/java/com/poly/demo/repository/VoucherRepository.java
@@ -1,21 +1,14 @@
 package com.poly.demo.repository;
 
-import com.poly.demo.entity.Seat;
-import com.poly.demo.entity.Showtime;
-import com.poly.demo.entity.Ticket;
-import com.poly.demo.entity.TicketFood;
-import com.poly.demo.entity.TicketFoodId;
-import com.poly.demo.entity.User;
-import com.poly.demo.entity.Voucher;
+import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
+import com.poly.demo.entity.Voucher;
 
 @Repository
-public interface VoucherRepository extends JpaRepository<Voucher, Long> {
+public interface VoucherRepository extends JpaRepository<Voucher, Integer> {
 	Optional<Voucher> findByCode(String code);
 
 }
